Revert "fix: fix social auth"
