+++ .env.sample
@@ -1,4 +1,1 @@
 API_URL=https://your-teams-backend-probe-url.com
-GOOGLE_CLIENT_ID=
-GOOGLE_CLIENT_SECRET=
-AUTH_SECRET=
