+++ package.json
@@ -39,48 +39,46 @@
     "@react-email/tailwind": "^0.0.18",
     "@t3-oss/env-nextjs": "^0.10.1",
     "@types/jest-axe": "^3.5.9",
-    "axios": "^1.7.2",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
     "cookies-next": "^4.2.1",
-    "framer-motion": "^11.3.19",
+    "framer-motion": "^11.3.8",
     "input-otp": "^1.2.4",
     "jest-axe": "^9.0.0",
     "jose": "^5.6.3",
-    "lenis": "^1.1.8",
+    "lenis": "^1.1.6",
     "lucide-react": "^0.400.0",
     "moment": "^2.30.1",
     "next": "14.2.5",
-    "next-auth": "5.0.0-beta.20",
+    "next-auth": "5.0.0-beta.19",
     "next-nprogress-bar": "^2.3.13",
-    "react": "^18.3.1",
+    "react": "^18",
     "react-day-picker": "^8.10.1",
-    "react-dom": "^18.3.1",
+    "react-dom": "^18",
     "react-email": "2.1.5",
     "react-hook-form": "^7.52.1",
-    "react-icons": "^5.2.1",
     "react-paginate": "^8.2.0",
     "recharts": "^2.12.7",
     "sharp": "^0.33.4",
-    "swiper": "^11.1.8",
-    "tailwind-merge": "^2.4.0",
+    "swiper": "^11.1.5",
+    "tailwind-merge": "^2.3.0",
     "tailwindcss-animate": "^1.0.7",
     "zod": "^3.23.8",
     "zustand": "^4.5.4"
   },
   "devDependencies": {
-    "@ianvs/prettier-plugin-sort-imports": "^4.3.1",
+    "@ianvs/prettier-plugin-sort-imports": "^4.3.0",
     "@mertasan/tailwindcss-variables": "^2.7.0",
-    "@testing-library/dom": "^10.4.0",
-    "@testing-library/jest-dom": "^6.4.8",
+    "@testing-library/dom": "^10.3.1",
+    "@testing-library/jest-dom": "^6.4.6",
     "@testing-library/react": "^16.0.0",
     "@testing-library/user-event": "^14.5.2",
-    "@types/node": "^20.14.13",
+    "@types/node": "^20.14.12",
     "@types/prop-types": "^15.7.12",
     "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@typescript-eslint/eslint-plugin": "^7.18.0",
-    "@typescript-eslint/parser": "^7.18.0",
+    "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^7.15.0",
+    "@typescript-eslint/parser": "^7.15.0",
     "@vitejs/plugin-react": "^4.3.1",
     "@vitest/coverage-istanbul": "^1.6.0",
     "@vitest/coverage-v8": "^1.6.0",
@@ -88,18 +86,18 @@
     "eslint": "^8.57.0",
     "eslint-config-next": "14.2.4",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
-    "eslint-plugin-react": "^7.35.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.34.4",
     "eslint-plugin-react-hooks": "^4.6.2",
     "eslint-plugin-testing-library": "^6.2.2",
     "eslint-plugin-unicorn": "^54.0.0",
     "eslint-plugin-unused-imports": "^3.2.0",
     "eslint-plugin-vitest": "^0.4.1",
-    "jsdom": "^24.1.1",
-    "postcss": "^8.4.40",
-    "prettier": "^3.3.3",
+    "jsdom": "^24.1.0",
+    "postcss": "^8",
+    "prettier": "^3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.5",
-    "tailwindcss": "^3.4.7",
+    "tailwindcss": "^3.4.1",
     "typescript": "^5.5.4",
     "vitest": "^1.6.0"
   }
