--- src/app/api/auth/[...nextauth]/route.ts
@@ -1,2 +1,0 @@
-export { GET, POST } from "~/auth";
-export const runtime = "edge";
