--- src/app/dashboard/(user-dashboard)/dashboard/page.tsx
@@ -1,5 +1,0 @@
-const page = () => {
-  return <div>page</div>;
-};
-
-export default page;
