+++ src/lib/auth.ts
@@ -1,0 +1,7 @@
+import NextAuth from "next-auth";
+
+const authOptions = NextAuth({
+  providers: [],
+});
+
+export const { handlers, signIn, signOut, auth } = authOptions;
