--- src/types/index.d.ts
@@ -1,8 +1,0 @@
-export interface IUser {
-  name: string;
-  email: string;
-  id: string;
-  role: "user" | "admin" | "super-admin" | "string";
-  access_token: string;
-  image: string;
-}
